Assistant checkpoint: Update token page layout to three columns with vertical scroll
Assistant generated file changes:
- client/src/pages/pumpfun-vision.tsx: Update layout to three columns with vertical scrolling
- client/src/components/TokenCard.tsx: Update TokenCard to be more compact, Update image size in TokenCard

---

User prompt:

can we make the design of the token page look like this. We want to keep all our visuals and looks but we want to make 3 separate parts for the tokens and display them way smaller as shown here. We want new token section left middle will be the about to graduate and right will be the graduated section. The tokens will show up and go down as the next comes instead of how they show up and go right side as they do now. Any questions ?